Remove anotações de precisão desnecessárias

+++ EP1/cholesky.f95
@@ -4,7 +4,7 @@ implicit none
 contains
   function cholcol(n, A)
     integer, intent(in) :: n
-    real*8, intent(inout) :: A(:, :)
+    real, intent(inout) :: A(:, :)
     integer :: cholcol
 
     cholcol = n
